DERBY-1330 (partial) Currently, in case of revoke privilege, DataDictionary.addRemovePermissionsDescriptor expects it's callers to set the permission descriptor's uuid so that DependencyManager can be invoked by addRemovePermissionsDescriptor to send REVOKE_PRIVILEGE action to permission descriptor's dependents (sending of REVOKE_PRIVILEGE work has not been finished yet. This patch is in preparation of that work). In order to set the uuid, the callers have to goto system tables to find the uuid and then set the permission descriptor's uuid.
It will be less error prone if DataDictionary.addRemovePermissionsDescriptor did the setting of uuid of
permission descriptor, rather than having every caller set the uuid correctly.
DataDictionary.addRemovePermissionsDescriptor method has to goto the permission system tables anyways
and has the uuid information handy and hence it will be more efficient for it to set the uuid of the
permission descriptor. This is inline with the patch that was commited some time back so that resetting
of the uuid happened in addRemovePermissionsDescriptor method rather than expecting the callers to do that.

In order to do this, I have added an abstract method to PermissionsCatalogRowFactory called setUUIDOfThePassedDescriptor.
 This method will set the uuid of the passed permission descriptor to same value as the row corresponding to the
permission system table which is also passed as a parameter. This method will be called by
DataDictionary.addRemovePermissionsDescriptor. 

Patch submitted by Mamta Satoor.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@423949 13f79535-47bb-0310-9956-ffa450edef68