DERBY-1248
This problem was caused because the backup of the database did not get all
the log records describing the changes to the data files that it got.
The problem is that the online backup reads page "through" the cache, rather 
than from disk so it is picking up "dirty" pages where the associated log 
records have not yet been flushed to disk. 
The backup of the log file on the other hand goes directly from disk and thus 
sometimes in this concurrent test misses some of the log records in the 
unflushed part of the log file. The fix is to flush the log file up to the
end at the end of the backup process after the data files have been copied
and before we copy the last of the log files to the backup.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@430396 13f79535-47bb-0310-9956-ffa450edef68