DERBY-688: Enhancements to XML functionality toward XPath and XQuery support
This revision contains d688_phase4_v2.patch.

This patch was contributed by Army Brown (qozinx@gmail.com).

The phase 4 patch, d688_phase4_v2.patch, adds some additional restrictions to
the ways in which XML values can be used. In particular:

  1. XML types cannot be used in CREATE PROCEDURE or CREATE FUNCTION statements.

  2. XML types cannot be used in import/export functions.

  3. XML types cannot be declared as columns in a global temp table.

I admit that I'm a bit fuzzy as to *why* these restrictions need to be in
place, but these are the restrictions that apply to the other "long" datatypes
in Derby (LOBs, LONG VARCHAR) so I'm enforcing them for XML, as well, to be
safe. It'll be easier to remove these restrictions in the future than it
will be to block them after users have potentially been relying on them.

One final restriction added by this patch is as follows: if a parameter is
used for the operand to an XMLPARSE operation, the parameter must have an
explicit cast to a character string type. I've put this restriction in place
because, based on my reading of the spec, this is required for SQL/XML[2006]
conformance. Further explanation can be found in comments for the relevant code.

d688_phase4_v2.patch also adds some simple test cases for each of these
restrictions to the lang/xml_general.sql test, with the corresponding master
updates.

And finally, the phase 4 patch includes two new error messages: one for the
XMLPARSE restriction, and one for missing XML classes, which is actually for
phase 5 but I included it in the phase 4 patch so that the two patches can
be applied sequentially (phase 4 then phase 5) without conflict. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@430626 13f79535-47bb-0310-9956-ffa450edef68