DERBY-415 : sysinfo with -cp client should not complain about DB2 Driver
This change modifies the behavior of the "-cp client" argument to the
sysinfo tool. Formerly, this argument checked for both the DerbyNetClient
and the DB2 JCC driver, and complained if both drivers were not present.

Increasingly, users of Derby are using just the DerbyNetClient, and the use
of the JCC driver is less common, so it makes sense that "-cp client" should
focus on the DerbyNetClient, and the DB2 JCC driver can be treated separately.

So, "-cp client" now only checks for the DerbyNetClient, and a new argument
"-cp db2driver" is added to check for the DB2 JCC driver. The new behavior is:
   java sysinfo -cp: checks all the various components of the classpath
   java sysinfo -cp client: just checks the network client
   java sysinfo -cp db2driver: just checks the JCC driver



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@431741 13f79535-47bb-0310-9956-ffa450edef68