DERBY-1473: Add cut-off and truncation logic to streaming classes in the embedded driver
This patch adds the capability to handle length less streams
properly. Such streams are now capped at the maximum length of the
column it is inserted into. If the stream is longer than the limit, a
DerbyIOException (subclass of IOException) is thrown, causing a
statement severity StandardException to be thrown. The exception
handling system takes care of cleaning up.

If a stream is shorter or longer than the specified length, an
exception is thrown as required by JDBC 3.0.

Contributed by Kristian Waagan.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@437127 13f79535-47bb-0310-9956-ffa450edef68