DERBY-1810
Bumping the time to wait for server to start in this test.  In my environment
this test is failing consistently (and bad error handling in the test then 
causes this test to hang forever).  Bumping the timeout so far has made it 
pass (I tried it 10 times), where before it failed 10 times.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@439702 13f79535-47bb-0310-9956-ffa450edef68