DERBY-1756 patch Derby1756.2.diff.txt contributed by Sunitha Kambhampati
with derby-962 changes, if client jvm supports EUSRIDPWD then the client would 
use EUSRIDPWD as the security mechanism. But it is possible that the server jvm
might not support EUSRIDPWD and the connection can fail.
When DERBY-1517, DERBY-1755 is fixed, there might be a way to use EUSRIDPWD 
when both client and server vm's have support for it.

This patch does the following:
o Do not use EUSRIDPWD as the default security mechanism even if the client vm can support it.
o Fix comments in testSecMec.java.
o Existing tests in testSecMec.java cover this codepath and the master file 
output reflects the changes made. Note, only the ibm14 client master file has 
changed since only ibm141 and greater jvms come with jce that can support 
eusridpwd.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@439775 13f79535-47bb-0310-9956-ffa450edef68