DERBY-1491: Provide ALTER TABLE functionality to change a column's default value
This change contains new test cases in the altertable.sql test which explore
the behavior of the ALTER TABLE ALTER COLUMN statement to alter a column's
defaults. Note that this patch only includes new tests; the current product
already supports the functionality to change a column's default value. The
code to support this was a side effect of the DERBY-119 changes, which wired up
the already-existing parser routines to the new ALTER TABLE t ALTER COLUMN c
statement.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@440550 13f79535-47bb-0310-9956-ffa450edef68