DERBY-1729, contributed by Yip Ng committing derby1729-trunk-diff03.txt patch.
The GrantNode and RevokeNode should have derived from DDLStatementNode instead
of MiscellaneousStatementNode. Subclassing DDLStatementNode will generate a
call to GenericResultSetFactory's getDDLResultSet() in the activation class
and invokes the GenericAuthorizer's authorize() method with the proper
parameters to enforce the correct semantics.

public ResultSet getDDLResultSet (Activation activation) throws StandardExceptio
n
{
        getAuthorizer(activation).authorize(activation, Authorizer.SQL_DDL_OP);
            return getMiscResultSet( activation);
}

Also adding an additional sql file for derbylang. The reason I didn't include
this in grantRevokeDDL.sql is because of name collision and this testcase is
one of the many additonal grant/revoke tests that I wrote and I'll like to
append the rest of those testcases to this file(grantRevokeDDL2.sql) when I
submit my patch for DERBY-1736.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@441140 13f79535-47bb-0310-9956-ffa450edef68