DERBY-1618: (store/BootAllTest.junit fails on with jdk13 based jvms)
Patch Contributed by Myrna van Lunteren.

On jdk131, in the situation invoked by the testSettingBootAllPropertyWithHomePropertySet() 
test in BootAllTest.java, when the DriverPropertyInfo.choices were accessed.
the problem was because in BaseMonitor.java there were no elements found resulting from
provider.getBootTimeServices(). 

underlying real issue was  database directory was not getting set to the
correct path in the storage factory (DirStorageFactory.java), when all the 
database are loaded (derby.system.bootAll=true) and system home is set. 

This patch correctly sets the database directory in the storage factory, in the
above mentioed scenario.  

Modified the test lang/GetPropertyInfoTest.java to access the choices,
description, and required fields, but it didn't actually duplicate the problem
in BootAllTest.java. But as that's in that test I thought it was unnecessary to
add more.





git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@443038 13f79535-47bb-0310-9956-ffa450edef68