DERBY-1589: CREATE TABLE throws NPE after DROPs and REVOKEs
FKConstraintDefinitionNode, which is the compiler implementation code
for the "foreign key ... references ... " constraint in the compiler,
is not properly registering that the statement containing the
foreign key constraint has a dependency on the table being referenced. Due
to the lack of the dependency, when the referenced table (t11ConstraintTest)
is dropped, the statement is not invalidated. 

This patch changes the compiler so that it creates a dependency from
a statement with a FOREIGN KEY clause in it to the table which is
referenced by that clause. The patch also contains a simple regression
test, based on the original bug.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@447644 13f79535-47bb-0310-9956-ffa450edef68