DERBY-147: ERROR 42x79 if specify same column twice and use ORDER BY
This patch contributed by Bernd Ruehlicke (apache@priwall.org)

This patch changes the getOrderByColumn method in ResultColumnList so that
it uses ResultColumn.isEquivalent to determine whether the column specified
in the ORDER BY clause is ambiguous or not. It is ok to select the ORDER BY
column multiple times, so long as all the occurrences are equivalent. If there
is an ambiguity, the statement is rejected and the user must reword it to
clarify which column is to be used for ordering the results.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@447877 13f79535-47bb-0310-9956-ffa450edef68