DERBY-1304: DROP VIEW does not always completely remove view
This patch was contributed by Yip Ng (yipng168@gmail.com)

The problem is that DROP VIEW's binding logic is using its superclass
QueryTreeNode's bind() method which just simply returns the object itself.
So there is no actual binding done in DROP VIEW at all! It should have
created a dependency on the statement, so when its associated descriptor
gets dropped, the statement can be invalidated accordingly. This explains
why the second DROP VIEW fails since it is not invalidated by the system.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@448026 13f79535-47bb-0310-9956-ffa450edef68